// web/src/api/generated/models/DailyQuotePreferencePatch.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DailyQuotePreferencePatch
 */
export interface DailyQuotePreferencePatch {
    /**
     * 
     * @type {boolean}
     * @memberof DailyQuotePreferencePatch
     */
    enabled?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof DailyQuotePreferencePatch
     */
    locale?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof DailyQuotePreferencePatch
     */
    selectedCategoryIds?: Array<string> | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof DailyQuotePreferencePatch
     */
    selectedSourceIds?: Array<string> | null;
}

/**
 * Check if a given object implements the DailyQuotePreferencePatch interface.
 */
export function instanceOfDailyQuotePreferencePatch(value: object): value is DailyQuotePreferencePatch {
    return true;
}

export function DailyQuotePreferencePatchFromJSON(json: any): DailyQuotePreferencePatch {
    return DailyQuotePreferencePatchFromJSONTyped(json, false);
}

export function DailyQuotePreferencePatchFromJSONTyped(json: any, ignoreDiscriminator: boolean): DailyQuotePreferencePatch {
    if (json == null) {
        return json;
    }
    return {
        
        'enabled': json['enabled'] === undefined ? undefined : json['enabled'] === null ? null : json['enabled'],
        'locale': json['locale'] === undefined ? undefined : json['locale'] === null ? null : json['locale'],
        'selectedCategoryIds': json['selectedCategoryIds'] === undefined ? undefined : json['selectedCategoryIds'] === null ? null : json['selectedCategoryIds'],
        'selectedSourceIds': json['selectedSourceIds'] === undefined ? undefined : json['selectedSourceIds'] === null ? null : json['selectedSourceIds'],
    };
}

export function DailyQuotePreferencePatchToJSON(json: any): DailyQuotePreferencePatch {
    return DailyQuotePreferencePatchToJSONTyped(json, false);
}

export function DailyQuotePreferencePatchToJSONTyped(value?: DailyQuotePreferencePatch | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'enabled': value['enabled'],
        'locale': value['locale'],
        'selectedCategoryIds': value['selectedCategoryIds'],
        'selectedSourceIds': value['selectedSourceIds'],
    };
}